Some changes.

+++ 1503-StructByLightning-LostSamurai-StrawManProject/Lost Samurai/Last Samurai/TileSystem.cpp
@@ -96,6 +96,8 @@ TileSystem::~TileSystem()
 void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 
 {
+	//Game::GetInstance()->GetFont().Draw("Loading", { Game::GetInstance()->GetScreenSize().width / 2, Game::GetInstance()->GetScreenSize().height / 2 }, 1, { 255, 255, 255, 255 });
+
 	//TileSystem* Tileload = new TileSystem();
 	m_Grid = new Grid();
 	//Map = new Tile;
@@ -153,7 +155,7 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 			m_Tile->m_EventString = tile->Attribute("TriggerEvents");
 		}
 		tile->Attribute("Traps", &m_Tile->Traps);
-		if (m_Tile->Traps == 1 )
+		if (m_Tile->Traps == 1)
 		{
 			m_Tile->TrapNames = tile->Attribute("TrapNames");
 		}
@@ -292,7 +294,7 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 				Rect->right = Rect->left + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileWidth;
 				Rect->bottom = Rect->top + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileHeight;
 				Traps["Spikes"].push_back(Rect);
-				
+
 			}
 			if (Map[m_Tile->PositionX][m_Tile->PositionY]->TrapNames == "Smashing")
 			{
@@ -306,7 +308,7 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 				Rect->right = Rect->left + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileWidth;
 				Rect->bottom = Rect->top + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileHeight;
 				Traps["Smashing"].push_back(Rect);
-				
+
 
 
 			}
@@ -324,7 +326,7 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 				Rect->right = Rect->left + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileWidth;
 				Rect->bottom = Rect->top + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileHeight;
 				Traps["Rocks"].push_back(Rect);
-				
+
 
 			}
 
@@ -340,7 +342,7 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 				Rect->right = Rect->left + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileWidth;
 				Rect->bottom = Rect->top + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileHeight;
 				Traps["darts"].push_back(Rect);
-				
+
 
 
 			}
@@ -357,7 +359,7 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 				Rect->right = Rect->left + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileWidth;
 				Rect->bottom = Rect->top + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileHeight;
 				Traps["Cannon"].push_back(Rect);
-				
+
 			}
 
 			if (Map[m_Tile->PositionX][m_Tile->PositionY]->TrapNames == "PopSpikes")
@@ -372,7 +374,7 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 				Rect->right = Rect->left + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileWidth;
 				Rect->bottom = Rect->top + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileHeight;
 				Traps["PopSpikes"].push_back(Rect);
-				
+
 			}
 
 			if (Map[m_Tile->PositionX][m_Tile->PositionY]->TrapNames == "Doors")
@@ -387,13 +389,13 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 				Rect->right = Rect->left + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileWidth;
 				Rect->bottom = Rect->top + Map[m_Tile->PositionX][m_Tile->PositionY]->m_TileHeight;
 				Traps["Doors"].push_back(Rect);
-				
+
 			}
 
-			
-			
-			
-		
+
+
+
+
 
 		}
 
@@ -408,7 +410,7 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 		}
 	}
 #pragma endregion
-	tilecollison = tilecollison->NextSiblingElement("Rect"); 
+	tilecollison = tilecollison->NextSiblingElement("Rect");
 	//tile = rect->FirstChildElement();
 	tile = tilecollison->FirstChildElement();
 	while (tile != nullptr)
@@ -424,14 +426,14 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 		tile->Attribute("CollisionRectRight", &Right);
 		tile->Attribute("CollisionRectBottom", &Bottom);
 		SGD::Rectangle rect1;
-		rect1.left = Left;
-		rect1.top = Top;
-		rect1.bottom = Bottom;
-		rect1.right = Right;
-		rect->left = rect1.left;// -Game::GetInstance()->GetCameraPosition().x / 2;
-		rect->top = rect1.top;// -Game::GetInstance()->GetCameraPosition().y / 2;
-		rect->right = rect1.left + rect1.ComputeWidth();// -Game::GetInstance()->GetCameraPosition().x / 2;
-		rect->bottom = rect1.top + rect1.ComputeHeight();// -Game::GetInstance()->GetCameraPosition().y / 2;
+		rect1.left = (float)Left;
+		rect1.top = (float)Top;
+		rect1.bottom = (float)Bottom;
+		rect1.right = (float)Right;
+		rect->left = (float)rect1.left;// -Game::GetInstance()->GetCameraPosition().x / 2;
+		rect->top = (float)rect1.top;// -Game::GetInstance()->GetCameraPosition().y / 2;
+		rect->right = (float)rect1.left + rect1.ComputeWidth();// -Game::GetInstance()->GetCameraPosition().x / 2;
+		rect->bottom = (float)rect1.top + rect1.ComputeHeight();// -Game::GetInstance()->GetCameraPosition().y / 2;
 		m_Tile->SetRect(*rect);
 		m_Tile->SetPosition(SGD::Point{ (float)rect->left, (float)rect->top });
 		m_CollisionRect.push_back(m_Tile);
@@ -439,13 +441,13 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 
 		tile = tile->NextSiblingElement();
 	}
-	
+
 
 	tilecollison = tilecollison->NextSiblingElement("Ledge");
 	tile = tilecollison->FirstChildElement();
 	while (tile != nullptr)
 	{
-		
+
 		SGD::Rectangle* rect = new SGD::Rectangle();
 		int Left;
 		int Top;
@@ -456,17 +458,17 @@ void TileSystem::LoadTileXml(Father* _father, Son* _Son)
 		tile->Attribute("CollisionRectRight", &Right);
 		tile->Attribute("CollisionRectBottom", &Bottom);
 		SGD::Rectangle rect1;
-		rect1.left = Left;
-		rect1.top = Top;
-		rect1.bottom = Bottom;
-		rect1.right = Right;
+		rect1.left = (float)Left;
+		rect1.top = (float)Top;
+		rect1.bottom = (float)Bottom;
+		rect1.right = (float)Right;
 		rect->left = rect1.left;// -Game::GetInstance()->GetCameraPosition().x / 2;
 		rect->top = rect1.top;// -Game::GetInstance()->GetCameraPosition().y / 2;
 		rect->right = rect1.left + rect1.ComputeWidth();// -Game::GetInstance()->GetCameraPosition().x / 2;
 		rect->bottom = rect1.top + rect1.ComputeHeight();// -Game::GetInstance()->GetCameraPosition().y / 2;
 		Traps["Ledges"].push_back(rect);
 		//delete m_Tile;
-		
+
 
 		tile = tile->NextSiblingElement();
 	}
